Centered quality badge. Extracted strings.
